Add: Weapon factories.

+++ src/items/mod.rs
@@ -22,5 +22,5 @@ pub use item_display_system::ItemDisplaySystem;
 pub use light_source::{LightSource, LightSourceSystem};
 pub use transfer_item::TransferItem;
 
-pub static ITEM_ICON_INDEX_TORCH: usize = 0;
-pub static ITEM_ICON_INDEX_WEAPON: usize = 1;
+pub static ITEM_ICON_INDEX_TORCH: u32 = 0;
+pub static ITEM_ICON_INDEX_WEAPON: u32 = 1;
